move quick-merge to utls
